Move comparable.ts to src/common
